requirements.txt
appdirs==1.4.4
apturl==0.5.2
asgiref==3.5.2
backports.zoneinfo==0.2.1
baron==0.10.1
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==1.0.0
dj-email-url==1.0.6
Django==4.1.3
django-cache-url==3.4.2
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-environs==0.1.1
django-js-asset==2.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
duplicity==0.8.12.0
emoji==1.6.1
entrypoints==0.3
environs==9.5.0
fasteners==0.14.1
filelock==3.6.0
future==0.18.2
httplib2==0.14.0
idna==2.8
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libretranslatepy==2.1.1
lockfile==0.12.2
louis==3.12.0
lxml==4.7.1
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
marshmallow==3.19.0
monotonic==1.5
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
packaging==21.3
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
platformdirs==2.5.1
protobuf==3.6.1
psutil==5.5.1
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
pyjokes==0.6.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.0.9
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==0.21.0
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
redbaron==0.9.2
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
rply==0.7.8
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sqlparse==0.2.4
systemd-python==234
translate==3.6.1
ubuntu-advantage-tools==27.9
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.13.2
wadllib==1.3.3
whitenoise==6.2.0
xkit==0.0.0
